Warnings cleanup

+++ ATK/Delay/HadamardMixture.h
@@ -23,7 +23,7 @@ namespace ATK
     static const unsigned int nb_channels = 1 << order;
     typedef DataType_ DataType;
     /// Gain factor to take into account in the feedback loop due to the instability of the mixture
-    static constexpr double gain_factor = 1;
+    static constexpr float gain_factor = 1.f;
 
     class MixtureImpl;
